pyproject.toml
[project]
name = "paperatlas"
version = "0.1.0"
description = "Concept-centric research paper recommendation system"
requires-python = ">=3.10"

dependencies = [
  "torch",
  "transformers",
  "sentence-transformers",
  "faiss-cpu",
  "numpy",
  "scipy",
  "pandas",
  "scikit-learn",
  "pydantic>=2.0",
  "pyyaml",
  "fastapi",
  "uvicorn",
  "httpx",
  "neo4j",
  "mysql-connector-python",
  "pymupdf"
]

[tool.uv]
dev-dependencies = [
  "pytest",
  "black",
  "ruff",
  "ipykernel"
]
